opal: run clang-format on all of opal
Some edits had to be done by hand as opal_datatype_internal.h is incorrectly
detected as objective-c. This happens because the OPAL_DATATYPE_INIT_PTYPES_ARRAY
macro seems to trip up clang-format. I intend to fix this bug upstream but
for now it was fine to delete this definition, format the code, then re-add
it to match the style guidelines.

Signed-off-by: Nathan Hjelm <hjelmn@cs.unm.edu>