clang-format-ompi: do not process fake header files
Open MPI has some "header" files that are really fotran includes but not named
with the expected extension (these are non-standard and sometimes named with .fi
for fortran include). They are not C header files so the clang-format script should
not attempt to format these files. The result is a mess and will fail to compile.

Open MPI should rename these files but for now they all have an emacs mode line
we can use to filter them out.

Signed-off-by: Nathan Hjelm <hjelmn@google.com>