Reenable the heterogeneous support.
This commit fixes the support for heterogeneous environments and
specifically for external32. The root cause was that during the datatype
optimization process types that are contiguous in memory are collapsed
together in order to decrease the number of conversion (or memcpy)
function calls. The resulting type however, does not have the same
conversion rules as the types it replaced, leading to an incorrect (or
absent) conversion in some cases. This patch marks the datatypes where
types have been collapsed during the optimization process with a flag,
allowing the convertor to detect if the optimized type can be used in
heterogeneous setups.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>
(cherry picked from commit 73d64cb97b48c4c6860aa0d49af9705349d6f64b)