Add a soft "breakpoint" capability
Provide the ability to define multiple "break points"
in the OMPI code base controlled by an "OMPI_BREAKPOINT"
envar. If the envar is set and is equal to the defined
breakpoint tag, and if a debugger has set the
PMIX_DEBUG_STOP_IN_APP directive, then we will stop at
that point, issue the PMIX_READY_FOR_DEBUG event, and
hold until released by the debugger.

This can be placed anywhere in the code by simply
including the "ompi_rte_breakpoint("tag")" code,
where "tag" is the string you will set OMPI_BREAKPOINT
equal to when you want that breakpoint to be active.

Signed-off-by: Ralph Castain <rhc@pmix.org>