Long live MPI_LONG and MPI_UNSIGNED_LONG
The only 2 types that have an external32 representation with a different
size than most current architectures, and as a result are more
challenging to handle.

This patch also brings back the support for packing and unpacking to and
from external32 for all datatypes and does a little cleaning in the
datatype API and comments.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>
(cherry picked from commit 4e56e83d524ff7e425afb8b15e9aeb04235c3189)