threads: Fix C++/Fortran pthread tests
Commit 38e53f3 introduced a bug in the checks for pthreads when
using C++ or Fortran, resulting in -pthread not being added
to the wrapper compiler flags.  This (on some platforms) lead
to link errors in multi-threaded applications.  Fix the bug by
not declaring success after the check for whether or not pthreads
works without additional libraries; this was overriding the
value set in the plain check.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>