threads: Removed unneeded Linux Threads check
Remove the "threads have different pids" check, which was really
a check for the old LinuxThreads behavior, as it is no longer
used.  In addition to NPTL being released 17 years ago and the
default basically everywhere, nothing in Open MPI uses the
macro set and this is just dead code at this point.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>