Update PMIx and PRRTE pointers
Fully implement the PMIx_Connect operation in support
of MPI connect/accept. At the end of the "connect" op,
every process has access to not only the job-level info
for all participating jobs, but also all modex info
provided by all participants. In addition, "connect"
returns a context ID that can be used as the communicator
ID. This value is returned under the first proc in
the array of procs passed to PMIx_Connect using the
PMIX_GROUP_CONTEXT_ID key.

If the first proc has wildcard rank, then the context ID
will be under rank=0 of that job.

Signed-off-by: Ralph Castain <rhc@pmix.org>