libevent: Link against libevent_core, not libevent
In the change to use 3rd-party packaging, a mistake was made and
we started linking against the old-style libevent.* instead of
the new-style libevent_core.*.  This caused errors on OS X, where
PMIx linked against libevent_core and OMPI linked against
libevent.  This patch changes the behavior back to linking
against libevent_core.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>