build: Avoid accidental calls to AC_PROG_CC
Reorder some of the early configure tests to avoid
accidently calling AC_PROG_CC before the C compiler
section, which will also ensure those tests are run
with the same option flags as the rest of the tests.

Reorder the wrapper compiler initialization, as it
depended on these accidental calls to AC_PROG_CC.
To make the late call to the wrapper compiler
code work, save the "bare" CC (before Autoconf
macros add stuff to it) in the env variable
OPAL_CC, similar to the AC_DEFINE of the same
name.

To keep this from happening again, cause an Autoconf
error if Autoconf expands AC_PROG_CC before the
call to OPAL_SETUP_CC, and put in some AC_REQUIRE
statements to better state dependencies.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>