[mpich romio c4b5762] romio flatten has flat->count consistency problems
> Pulled in from mpich romio, branch "main".
> Their commit message is below.
>
> This is part of a batch of commits from the
> following set of PRs:
>  *  https://github.com/pmodels/mpich/pull/4943
>     -- darray fix which contains a flatten fix
>         73a3eba
>         c4b5762
>  *  https://github.com/pmodels/mpich/pull/4995
>     -- write strided fix
>         bbb5210
>  *  https://github.com/pmodels/mpich/pull/5100
>     -- build fix for -Wpedantic
>         ad0e435
>  *  https://github.com/pmodels/mpich/pull/5099
>     -- build fix, they had let file-system=...gpfs bit rot
>         e1d42af
>         313289a
>         83bbb82
>  *  https://github.com/pmodels/mpich/pull/5150
>     -- build fix, configure-related _GNU_SOURCE
>         a712b56
>         5a036e7
>  *  https://github.com/pmodels/mpich/pull/5184
>     -- build fix, continuation of _GNU_SOURCE fix
>         d97c4ee

ADIOI_Count_contiguous_blocks and ADIOI_Flatten are very similar functions,
but they diverge a lot around the edges and in degenerate cases.  In Spectrum
MPI I spent some time making them consistent with each other but
found that to be a losing battle.

So the approach I used here is to not have Count() be as definitive,
and rather let Flatten() have the last word on what the final flat->count
really is.  Eg Flatten's *curr_index is the real count.

The changes I made are

1. Fix a couple places in Flatten where *curr_index was updated out
of sync with what was actually being added to flat->indices[] and
flat->blocklens[].  That's one of the important book-keeping rules
Flatten should follow.  There were a couple places (when counts[] arrays
have 0s) where that wasn't the case (see the "nonzeroth" var in this
commit).

2. The main change I made was to reset flat->count based on
Flatten's curr_index.  This is because the divergence between the
two functions is too great to reliably fix.

3. A third change is just a safety valve, using a
flatlist_node_grow() function just in case the
Count function returns a smaller value than Flatten ends up
trying to write into the array, and related to this I
got rid of the various --(flat->count) lines, since that var now
represents the allocated size of the array, until right after
the Flatten function when it's reset to represent the data written
to the array like it used to be.

I don't think we even need ADIOI_Count_contiguous_blocks() anymore,
but I didn't remove it.

Signed-off-by: Mark Allen <markalle@us.ibm.com>
(cherry picked from commit 3e7f46b1c3fbe31c19e5ce0af2be29694b412378)