configury: do not use _Float16 if noisy build
Turns out that with more recent gcc compilers at least (9.1.0? and newer)
_Float16 is detected on aarch64, but by default we are compiling OMPI
with -pendantic.

This results in thousands and thousands of warnings of this type to be emitted
while building:

./../../opal/include/opal_config.h:2311:28: warning: ISO C does not support the ‘_Float16’ type [-Wpedantic]
 2311 | #define opal_short_float_t _Float16
      |                            ^~~~~~~~
../../../opal/include/opal_config_bottom.h:541:5: note: in expansion of macro ‘opal_short_float_t’
  541 |     opal_short_float_t real;
      |     ^~~~~~~~~~~~~~~~~~
../../../opal/include/opal_config.h:2311:28: warning: ISO C does not support the ‘_Float16’ type [-Wpedantic]
 2311 | #define opal_short_float_t _Float16
      |                            ^~~~~~~~
../../../opal/include/opal_config_bottom.h:542:5: note: in expansion of macro ‘opal_short_float_t’
  542 |     opal_short_float_t imag;

this patch adds a -Werror to CFLAGS before checking for _Float16 to fail the check
if warnings are emitted by the compiler.

Related to issue #8840

Signed-off-by: Howard Pritchard <howardp@lanl.gov>
(cherry picked from commit 9e2d9bd35b94942da67da14946528855e6de8965)