Fix a bug in the cherry pick checker
The check_cherry_pick function in .github/workflows/git-commit-checks.py
would not function propperly in the case where the commit being checked
is a cherry-pick from a commit that is on an unmerged pull request on
github. This has been resolved, so that the workflow will exit with a
failure unless the referenced commit has been successfully merged.

Signed-off-by: StevenGood154 <steve.good.154@gmail.com>
Signed-off-by: bwiseman77 <brett.wiseman@comcast.net>
Signed-off-by: cvankirk2 <christinenvankirk@gmail.com>