CI: Fix a bug in the cherry pick checker
The check_cherry_pick() function in .github/workflows/git-commit-checks.py
would incorrectly allow some cherry-pick messages reffering to commits that
do not exist in the base repository, specifially when the commit exists only
on a pull request that has not been merged. This has been resolved; the
workflow will exit with a failure unless the referenced commit exists in
the base repository.

Note that this approach only works because all commits are introduced to
the open-mpi/ompi through pull requests from forks.

Signed-off-by: Steven Good <steve.good.154@gmail.com>
Signed-off-by: Brett Wiseman <brett.wiseman@comcast.net>
Signed-off-by: Christine Van Kirk <christinenvankirk@gmail.com>