libevent: Do not force addition of linker flags
OPAL_CHECK_PACKAGE expects the libdir argument to be an empty string
unless the user specifically requested a specific libdir.  In that
case, we should always add a linker search flag for that directory,
as the user is specifying a directory location.  The external
libevent configure code was trying to find an external libevent
library and then set the libdir to OPAL_CHECK_PACKAGE as the
found libdir, but without all the smarts to not add unnecessary
search flags.  This resulted in a -L/usr/lib appearing early in
our explicit search paths if a user specified --with-libevent=/usr.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>
(cherry picked from commit 75e82db669c9b70ab6e8ae15de4d52a688e4b023)