atomics: add the --disable-cx16-atomics
Do not try the -mcx16 flag if --disable-cx16-atomics is used,
and prevent the generation of instructions that are not available
on all x86 platforms (such as Celeron N4000).

Always try to run a simple test to make sure the selected atomic
generate correct results.

Thanks Orion Poplawski for reporting this issue.

Refs. open-mpi/ompi#9022

Signed-off-by: Gilles Gouaillardet <gilles@rist.or.jp>