Create an mpirun dedicated executable
If we install the embedded versions or PRRTE and PMIx, then
OPAL_PREFIX needs to adjust the installed location for PMIx
as well as for PRRTE when launching via mpirun.

In mpirun: Detect OPAL_PREFIX and set the corresponding PRRTE and PMIx
prefixes. Take advantage of the simplified proxy detection
support in PRRTE. Deal with absolute path names

In MPI_Init: Detect OPAL_PREFIX and set the corresponding PMIx
prefix

Signed-off-by: Ralph Castain <rhc@pmix.org>