pmix3x: fix slurping of PMIx compiler/linker flags
Previously, after invoking PMIx's "configure" script, we grep'ed
PMIx's flags from its config.status.  Apparently, the values were
includes in a section of config.status that was actually an awk
script, and if the value got long enough, GNU Autoconf would use awk
line continuation syntax to span multiple lines.  This meant that Open
MPI's grep would basically get an incomplete value.

This commit changes to instead grep the value from a PMIx Makefile,
where the value will never be split across multiple lines.

Thanks to Erik Schnetter for reporting the issue.

This is not a cherry pick from master because the problem does not
exist on master due to the refactoring of how PMIx is configured and
built.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>