Make MPI_[Comm|File|Win]_get_info MPI 4.0 compliant
So far Open MPI has elected to not hand out values for keys
that were modified through MPI_*_set_info.
MPI 4.0 requires implementations to reflect such changes if the
the new values are supported by the implementation.
Hence, it is sufficient to simply duplicate the info object and
remove the handling of OPAL_INFO_SAVE_PREFIX from OPAL as there
are no other users of this infrastructure.
Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>