v5.0.x: common/ofi: Disable new monitor API until libfabric 1.14.0
There are known issues with the API in libfabric 1.13.0 which will guarantee
segfaults when used. These issues are fixed in libfabric 1.13.1, but we
do not have a way to detect which patch version of libfabric is used.
Thus, delay the usage of the API until the subsequent minor release.

Signed-off-by: William Zhang <wilzhang@amazon.com>
(cherry picked from commit 190feba613fd6fb621550b7f097fb2b5672a5b98)