configury: explicitly disable gcc < v4.8.1
NOTE: The default GCC that ships in RHEL is v4.8.5.  Cisco regularly
tests GCC v4.8.1, and it works fine.  We know that PRTE and PMIX do
not compile with RHEL 6's default GCC (v4.4.7).  We know that v4.8.x
works and 4.4.7 doesn't, so set the lower bound support for 4.8.1.

Specifically: abort configure if the compiler is GCC and the version
is < v4.8.1.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>