osc/rdma: use btl/self for self communication as last resort
ompi_osc_rdma_peer_btl_endpoint() is used to select btl and endpoint
to communicate with a peer.

This patch added a change to ompi_osc_rdma_peer_btl_endpoint() that:
if no btl/endpoint has been selected for self communication,
and if bml has the btl/self, then use btl/self for self communication.

It also made a change to ompi_osc_rdma_new_peer():

Currently if no btl can be found and the peer is self, the function
still continues. This patch made the function to fail in this case,
because the ability to do self communication is essential for osc/rdma.

Signed-off-by: Wei Zhang <wzam@amazon.com>