Fortran mpi module: remove interface names
Fortran is a deceptively complicated language.  It turns out that it
is permissable to pass arrays of many sizes and shapes to subroutines
that expect 1-dimensional array parameters.  For example, a C
programmer might look at the following subroutine declaration and
assume that it means that the ARRAY_OF_REQUESTS parameter is a
1-dimensional array of any length:

```
MPI_WAITALL(COUNT, ARRAY_OF_REQUESTS, ARRAY_OF_STATUSES, IERROR)
    INTEGER COUNT, ARRAY_OF_REQUESTS(*),
              ARRAY_OF_STATUSES(MPI_STATUS_SIZE, *), IERROR
```

However, rules of the Fortran language allow the compiler to
effectively morph non-1-dimensional array source parameters to match,
for example, the 1-dimensional ARRAY_OF_REQUESTS dummy parameter.
Sometimes this will be done by copy-in-copy-out; other times, the
compiler can just swizzle the types to match because the source array
data is already contiguous in memory.

Regardless, there are several rules about when this type swizzling can
happen, and there are differences in the rules between when the target
subroutine in question is in a named interface such as:

```
INTERFACE FOO
  SUBROUTINE FOO(arg)
    INTEGER ARG(*)
  END SUBROUTINE FOO
END INTERFACE
```

versus when the subroutine is in an unnamed interface such as:

```
INTERFACE
  SUBROUTINE FOO(arg)
    INTEGER ARG(*)
  END SUBROUTINE FOO
END INTERFACE
```

I will not pretend to be a Fortran language expert and won't quote the
language rules here.  The short version is that unnamed interfaces
allow more flexible type swizzling.  For example, a user can pass a
2-dimensional INTEGER array to MPI_WAITALL's ARRAY_OF_REQUESTS
argument, and -- assuming that the 2D array is contiguous in memory --
the compiler will make the 2D array type match the 1D
ARRAY_OF_REQUESTS argument.  Open MPI will get a contiguous chunk of
memory of integers, so it doesn't know/care.

The MPI standard does not address this issue at all.  I.e., it doesn't
say whether passing N-dimensional parameters to subroutines like
MPI_WAITALL are permissable or not.  It also does not specify whether
interfaces need to be used, or if interfaces are used, whether they
need to be named or unnamed -- these are all implementation decisions.
However, there has been a lot of discussion about this in the Fortran
group at the MPI Forum over the last week; there will almost certainly
be an errata to MPI-4.0 and/or some text changes in MPI-4.1 to clarify
the situation.  I'm not going to speculate on the final language that
will get accepted by the Forum, but it seems like since Open MPI's
implentation decision to use named interfaces is perhaps the most
restrictive in terms of compiler swizzling functionality, we should
probably ease those restrictions so as not to violate the Law of Least
Astonishment from the user's perspective (who would expect that the
Fortran compiler will be able to swizzle array sizes and shapes).

As such, this commit removes all interface names from the Fortran
"mpi" bindings module where there's only a single subroutine in the
interface, and that subroutine name exactly matches the interface
name.  Note that this specifically excludes the following routines,
which have multiple subroutines in the interface (for two different
types of Fortran pointers):

* MPI_Alloc_mem
* MPI_Win_allocate
* MPI_Win_allocate_shared
* MPI_Win_shared_query

Deleting the interface names *may* cause ABI implications in some
compilers.  I have not exhaustively tested Fortran compilers to know
if this is an issue or not.  The safest thing to do is to implement
this change for Open MPI v5.0.0 where we're breaking lots of ABI
things as compared to prior versions of Open MPI.

Note that this issue does *not* affect the mpi_f08 module.  Using
named interfaces were distinctly specified in the MPI standard.

Also, this doesn't affect mpif.h, because Open MPI doesn't declare
most interfaces/subroutines in mpif.h.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>
(cherry picked from commit 2798fd82e515aeb492f84fb10b493e9e25a9f00a)