ofi: Refactor MCA vars / fix output level
The ofi common refactoring patches broke handling of
the verbose flag because I forgot that the input
variable has to live longer than the function (sigh).
Move the output level variable to be a global variable.

Add error checking to the entire mca registration
path, since it was notably lacking.

Even before the refactor, we didn't handle re-setting
verbosity after registering synonyms properly.  This
patch probably isn't perfect in that regards, but
at least is *consistent* in what we do.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>