snprintf: ignore some compiler warnings
Compilers are fairly unhelpful when telling us that it's possible that
an snprintf() may truncate its copy if it can tell at compile time
that the length of the source strings may be longer than the
destination string.  We know!  That's why we're using snprintf()!

Instead of explicitly using opal_snprintf() to get around the warning,
use the gcc compile flag -Wformat-truncation=0 to disable this warning
when in picky mode.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>