build: Fix OPAL_CHECK_COMPILER_VERSION_ID escaping
Fix escapling and simplify handling of literals in the OPAL_CHECK_-
COMPILER_VERSION_ID macro.  The macros were under-escaped,, which
was not causing problems but was also wrong.  The use of m4_lower()
meant that the argument to the helper macros had to be a string
literal, so enforce that requirement explicitly and then remove
the evals that are unnecessary given the lack of shell expansion.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>