osc/rdma: add local leader's pid in shm file name to make it unique
In osc/rdma, each communicator will assign a rank as the local leader
on one node. The local leader will create a shm file, and other ranks
in the same communicator and on the same node will attach to the shm
file.

For osc/rdma to work properly, the shm file name must be unqiue for
each communicator/node. To achieve that, osc/rdma used the following
shm file name:

    osc_rdma.<node_name>.<job_id>.<comm_id>

However, this name format did not achieve the goal, because comm_id
is only unique from process level. It can happen that different
communicator have same comm_id, as long as they do not share process.

To address this issue, this patch added local leader's pid to shm
file name to make it unique.

Signed-off-by: Wei Zhang <wzam@amazon.com>