configury: String OPAL_BUILD_PLATFORM_COMPILER_VERSION_STR
Fixes #9494

Before this change, the `OPAL_BUILD_PLATFORM_COMPILER_VERSION_STR`
macro is defined as a non-string character sequence (no enclosing
double quotes) in `opal_config.h`.

This is problematic if it contains `//` because it is treated as
a start of comment. Also, if it contains `(` or `)`,
`PLATFORM_STRINGIFY(OPAL_BUILD_PLATFORM_COMPILER_VERSION_STR)`
will result in a compilation error. Probably this is the case of
the issue #9494 because some builds of Clang/LLVM 9 and 10 set
`__clang_version__` like below.

```
clang version 10.0.1 (https://github.com/llvm/llvm-project.git ef32c611aa214dea855364efd7ba451ec5ec3f74)
```

This commit changes `OPAL_BUILD_PLATFORM_COMPILER_VERSION_STR`
to a string (enclosing with double quotes).

I think this is a proper approach because it is consistent
with the `PLATFORM_COMPILER_VERSION_STR` macro, where
`opal/include/opal/opal_portable_platform_real.h` defines it
as a string.

Signed-off-by: KAWASHIMA Takahiro <t-kawashima@fujitsu.com>