build: Improve handling of compiler version string
Stop saving the build compiler name and version string during
configure and move that detection to build time.  Expanding
out the version string in particular was causing escaping issues
that will not exist when the string is only expanded at use.
We need to check the compiler id and version integers during
configure, so that we can define them in mpi.h to compare
against the running compiler at application (not OMPI) build
time, but the strings were not being used in those situations.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>