Enable mpirun to work out-of-the-box when using external elements
While working with the Fedora team to track down a problem, we found
that mpirun failed when configured with external PMIx/PRRTE due to
not finding the executable or libraries. Of course, it worked IF we
set the PATH and LD_LIBRARY_PATH correctly to include both external
packages - but it seemed weird to configure with those paths, and then
to have mpirun just fail.

Simple enough to use the configured paths to help mpirun find what
it was told to use, and it may reduce user surprise.

Also, correctly handle --with-prrte=external
Ensure that PRTE_PATH in this case correctly points to the
prterun executable, not prte. Extract the library path in
this case to setup LD_LIBRARY_PATH.

Signed-off-by: Ralph Castain <rhc@pmix.org>