osc/rdma : fence mechanism for btls do not support remote completion
Currently, fence is implemented by two steps:

First, waiting for RDMA operiatons to complete locally.
Second, call coll->barrier on the communicator.

This is correct only if selected BTL support remote completion.
Otherwise, it can happen that when coll->barrier() finished, the remote
side of the RDMA operation has not completed yet.

This patch implemented a different barrier mechanism, which is used
when any of the selected BTL does not support remote completion.

In which case, each process will post an atomic operation to every
peer to increase a counter on the peer through the selected BTL endpoint.
Though wait for its own counter to reach number of peers. This ensures
all previous RDMA operations have completed.

Signed-off-by: Wei Zhang <wzam@amazon.com>