ext: Provide post_config hook
Add a post-config hook for MPI extensions, similar to the same
functionality in the MCA system.  The purpose of the post-config
hook is to provie a macro which is executed after configure has
decided whether or not to build an extension.  This is useful for
AC_DEFINEs or AM_CONDITIONALs.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>
(cherry picked from commit 790f4a3ff31e8c2cf83dd22c709b6421db3dfff4)