mtl/ofi: call fi_setopt to state MPI p2p requirements for CUDA
Call fi_setopt after endpoint initialization to tell Libfabric that while
device peer to peer support should be enabled when possible for network
transfers, it is not required and copies may be used instead.

An MCA parameter needs to be added for this in the future so that users can
toggle this option.

Signed-off-by: Robert Wespetal <wesper@amazon.com>