osc/rdma: do NOT proceed if no endpoint can be found for a peer
Currently, ompi_osc_rdma_new_peer() can still proceed if no btl endpoint has
been found a new peer. This caused segfault when a process try
to communicated with the peer.

This patch change it to: if a btl endpoint cannot be found for
a peer, ompi_osc_rdma_new_peer() wil not proceed.

Signed-off-by: Wei Zhang <wzam@amazon.com>