btl/uct: set reasonable defaults
This commit updates the defaults for btl/uct to allow Mellanox HCAs (mlx4_0, and mlx5_0)
to allow osc/rdma to work out of the box if selected. The component does not allocate
any UCX resources until add_procs so this should not have any impact on the default
performance.

References #9580

Signed-off-by: Nathan Hjelm <hjelmn@google.com>