Replace PMIX/PRRTE macros to enable co-build mode
The PMIX/PRRTE team decided not to carry the patches to support building
PMIX and PRRTE as embedded packages (in Open MPI or others).  After
much discussion, we decided the best path forward for Open MPI was to
patch PMIX/PRRTE to provide the functionality we need.  This patch
provides the patching of PMIX/PRRTE to support building PMIX/PRRTE
in an embedded mode (with embedded libevent/hwloc, if needed).

Another patch is required to Open MPI to actually use this
infrastructure.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>