Do not run PMIX/PRRTE configure when not needed
If the user specifies an external PMIX or PRRTE (or disables
PRRTE), skip running the sub-configure script.  The advantage of
this change is that it gives the user a way out (an external build)
if there is an error in our configure logic for the sub projects (which
has recently happened).  The down side is that "make dist" is broken
whenever OMPI is configured against an external PMIX/PRRTE library or
PRRTE is disabled.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>
(cherry picked from commit 1fe8ca9b83e5abf3378875e9cb537a5f69276067)