Break "make dist" if a package isn't configured
With a previous commit, it is now possible that configure will not
run the PMIX or PRRTE configure scripts.  This means that "make dist"
will not recurse into the package's directory and those packages will
unexpectedly not be in the dist tarball.  This patch avoids that
situation by explicitly breaking "make dist" in those situations,
printing the package that caused the issue.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>