osc/rdma: Enforce required BTL remote completion
The RDMA OSC component requires remote completion for its
synchronization algorithms to work properly.  This is phase one of two
to handle BTLs which do not provide remote completion.

This patch requires "accelerated" BTLs to support remote completion.
Accelerated BTLs are those which can communicate with all peers in the
Window (so multiple BTLs per Window are not used), natively provide
the required put/get/atomic operations, and which can register suitable
memory for the window's communication.

Signed-off-by: Wei Zhang <wzam@amazon.com>
Signed-off-by: Brian Barrett <bbarrett@amazon.com>