Always wait to complete all receives.
Without waiting for the last receive before going to the next iteration we
might overwrite data if the current left neighbor become the right at the next
iteration.

Start with an MPI_REQUEST_NULL request.
If the request is not NULL and the first rcounts is 0, the ompi_request_wait
will segfault.

Signed-off-by: George Bosilca <bosilca@icl.utk.edu>