Cleanup #ident warnings.
- runtime/opal_init.c:86: warning: #ident is a GCC extension
- runtime/ompi_mpi_init.c:123: warning: #ident is a GCC extension
- pinit_f.c:29: warning: #ident is a GCC extension

'#ident' was added at the request of Sun a long time ago, however it probably
is no longer important since modern packaging systems tend to
maintain good control of what files are associated with what packages.
It could also be used to detect capabilities in certain environments (such as
debuggers). That said - it's not clear that anyone is relying on this,
so it is probably safe to remove and clean up the warnings.

This commit is essentially a revert of:
https://github.com/open-mpi/ompi/commit/005652c9d41543494a25eab61fb10cdb0cec817b

If anyone wants this funtionality back, it is easy to put back.

Signed-off-by: Austen Lauria <awlauria@us.ibm.com>