Update BTL documentation
Add a more detailed description of the return value for btl_send,
adding information about the behavior when the call returns 1
(meaning the descriptor could be immediately completed).

Reformat the function documentation to follow doxygen convention,
including use of argument direction notation and putting the
param and return sections at the end of the description block.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>